services/product-service/requirements.txt
# Flask Core
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0

# API Documentation
flask-smorest==0.42.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Caching & Performance
Flask-Caching==2.1.0
redis==4.6.0

# Validation & Serialization
webargs==8.3.0
apispec==6.3.0

# Development & Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0
factory-boy==3.3.0
faker==19.6.2

# Code Quality
black==23.7.0
flake8==6.0.0
mypy==1.6.1
pre-commit==3.4.0

# Utilities
python-dotenv==1.0.0
click==8.1.7
Werkzeug==2.3.7
gunicorn==21.2.0

# Monitoring & Logging
structlog==23.1.0